requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
ca-certificates=2021.4.13=hecd8cb5_1
cairo=1.14.12=hc4e6be7_4
certifi=2020.12.5=py37hecd8cb5_0
cloudpickle=1.6.0=py_0
cycler=0.10.0=py37_0
cytoolz=0.11.0=py37haf1e3a3_0
dask-core=2021.5.0=pyhd3eb1b0_0
decorator=4.4.2=pyhd3eb1b0_0
expat=2.3.0=h23ab428_2
fontconfig=2.13.1=ha9ee91d_0
freetype=2.10.4=ha233b18_0
fribidi=1.0.10=haf1e3a3_0
fsspec=0.9.0=pyhd3eb1b0_0
gettext=0.21.0=h7535e17_0
glib=2.63.1=hd977a24_0
graphite2=1.3.14=h38d11af_0
graphviz=2.40.1=hefbbd9a_2
harfbuzz=1.8.8=hb8d4a28_0
icu=58.2=h0a44026_3
imageio=2.9.0=pyhd3eb1b0_0
intel-openmp=2021.2.0=hecd8cb5_564
jpeg=9b=he5867d9_2
kiwisolver=1.3.1=py37h23ab428_0
lcms2=2.12=hf1fd2bf_0
libcxx=10.0.0=1
libffi=3.2.1=h0a44026_1007
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libtiff=4.1.0=hcb84e12_1
libxml2=2.9.10=h7cdb67c_3
llvm-openmp=10.0.0=h28b9765_0
locket=0.2.1=py37hecd8cb5_1
lz4-c=1.9.3=h23ab428_0
matplotlib=3.3.4=py37hecd8cb5_0
matplotlib-base=3.3.4=py37h8b3ea08_0
mkl=2021.2.0=hecd8cb5_269
mkl-service=2.3.0=py37h9ed2024_1
mkl_fft=1.3.0=py37h4a7008c_2
mkl_random=1.2.1=py37hb2f4e1b_2
ncurses=6.2=h0a44026_1
networkx=2.5.1=pyhd3eb1b0_0
numpy=1.20.2=py37h4b4dc7a_0
numpy-base=1.20.2=py37he0bd621_0
olefile=0.46=py37_0
openssl=1.1.1k=h9ed2024_0
pango=1.42.4=h7e27002_1
partd=1.2.0=pyhd3eb1b0_0
pcre=8.44=hb1e8313_0
pillow=8.2.0=py37h5270095_0
pip=21.0.1=py37hecd8cb5_0
pixman=0.40.0=haf1e3a3_0
pygraphviz=1.3=py37h1de35cc_1
pyparsing=2.4.7=pyhd3eb1b0_0
python=3.7.7=hc70fcce_0_cpython
python-dateutil=2.8.1=pyhd3eb1b0_0
pywavelets=1.1.1=py37haf1e3a3_2
pyyaml=5.4.1=py37h9ed2024_1
readline=8.1=h9ed2024_0
scikit-image=0.18.1=py37hb2f4e1b_0
scipy=1.6.2=py37hd5f7400_1
setuptools=52.0.0=py37hecd8cb5_0
six=1.15.0=py37hecd8cb5_0
sqlite=3.35.4=hce871da_0
tifffile=2020.10.1=py37h0cf3a3e_2
tk=8.6.10=hb0a8c7a_0
toolz=0.11.1=pyhd3eb1b0_0
tornado=6.1=py37h9ed2024_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
yaml=0.2.5=haf1e3a3_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
